+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Added rounding for graphites from and to time range filters
   for very short absolute ranges (Issue #320)
 - Increased resolution for graphite datapoints (maxDataPoints), now equal to panel pixel width. (Closes #5)
-
+- Improvement to influxdb query editor, can now add where clause and alias (Issue #331, thanks @mavimo)
 
 # 1.5.3 (2014-04-17)
 - Add support for async scripted dashboards (Issue #274)